Change Ball.py directory
